Fix: Ansible sanity check doc-required-mismatch for argument 'filters'
Argument 'filters' in argument_spec was configured as not required, but was documented
as being required.

Altering the documentation to reflect code fixes this sanity check failure.

Example of exact error message being fixed:

```
ERROR: plugins/modules/import_keypair.py:0:0: doc-required-mismatch:
Argument 'filters' in argument_spec is not required, but is documented as being required
```