Fix: Ansible sanity check parameter-list-no-elements for argument 'paths'
Argument 'paths' in argument_spec was of type list but the type of
the elements in the list was not defined.

Adding `elements='str'` in code and documentation resolves this sanity check failure.

Exact error message being fixed:

```
ERROR: plugins/modules/validate_resource_files.py:0:0: parameter-list-no-elements:
Argument 'paths' in argument_spec defines type as list but elements is not defined
```