Fix: Only run sanity and unit tests with the python version in the venv
Only target the version of python3 in the activate virtual environment.
Do not target any other discovered python runtimes.

An update in ansible-test now triggers ansible-test to discover all available
python runtimes and attempt run for all discovered versions. This is not
our use case. We only want to test with the python runtime in the active
virtual environment.