Fix: Ansible sanity checks - Disabling of logging for ssh_key_path module arguments is not required
Only the path is logged and not the actual key content.

To disable we add the file and the sanity check, no-log-needed, to the
ignores file. There does not appear to be a more granular way of disabling
other than per file.

See https://docs.ansible.com/ansible/latest/dev_guide/testing/sanity/ignores.html
for more details.